+++ client/src/components/ButtonNotUser.js
@@ -1,0 +1,14 @@
+import { useState , useEffect} from "react";
+import { useHistory } from 'react-router-dom'
+
+function ButtonNotUser({ cartProductsArr, setCartProductsArr, currentUserID, currentUser, setCurrentUser,setCurrentOrder  ,currentOrder })
+{
+   
+    
+
+   
+    return(
+    <div><p>will need to put current order_product e.g. cart</p>Cart with user {currentUser}</div>
+    )
+  };
+export default ButtonNotUser;
