+++ tmp/pids/server.pid
@@ -1,0 +1,1 @@
+92242
